pyproject.toml
[project]
name = "gavmild"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apscheduler>=3.11.2",
    "email-validator>=2.3.0",
    "flask>=3.1.3",
    "flask-bcrypt>=1.0.1",
    "flask-login>=0.6.3",
    "flask-sqlalchemy>=3.1.1",
    "flask-wtf>=1.2.2",
    "gunicorn>=25.1.0",
    "openai>=2.28.0",
    "pymysql>=1.1.2",
    "python-telegram-bot>=22.7",
    "pytz>=2026.1.post1",
    "requests>=2.32.5",
    "ruff>=0.15.6",
    "user-agents>=2.2.0",
    "webauthn>=2.7.1",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-flask>=1.3.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
